Move menuitems to layouts folder
